requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# NFL data
nfl-data-py>=0.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipython>=8.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
